WIP: Draft for use of new variant system
This is a work in progress and relies on a not finialized version of
meta-pelux.

The point of this patch is to showcase one example on how to do the
uncommenting of local.conf and bblayers.conf which is required by the
suggested change to handle variants.

It requires new types of meta data to be sent to the CI, namely machine
and a list of features to enable. However, it does not propose a means
or format of doing so.